Substitute some patterns even when they use html entities
We treat HTML as first class support.
If there is a preprocessing stage generating a HTML content that is
supposed to be then passed to tipograph, some tools may escape
problematic characters (like " or >) with their HTML entity equivalents.
This breaks the capability of tipograph to discover the patterns that
are composed of these characters. This commit addresses this issue
partially.